Fifth Edition

+++ E-Commerce Website.BL/IRepositories/IServiceCart.cs
@@ -11,9 +11,9 @@ namespace E_Commerce_Website.BL.IRepositories
     {
         Task<List<CartUpdateViewModel>> GetAllUserCart(string Name);
         Task<string> AddToCart(CartViewModel cartViewModel, string name);
-        Task<double> totalprice(string name);
+        Task<decimal> totalprice(string name);
         Task<string> UpdateCartItem(CartUpdateViewModel cartUpdateViewModel);
-        Task<string> DeleteCartItem(int id,string name);
+        Task<string> DeleteCartItem(int id, string name);
         Task<string> DeleteAllCartItems(string name);
         Task<int> GetCartCount(string name);
     }
